Add iterator to bitboard that iterates over pieces that can move.
This is the same iterator type as the one used to iterate over all pieces,
but it's initialised with different bitboard.
New methods added to create this iterator:
can_move_begin() and can_move_end().
Also, method can_move_empty() added to check whether
there are no pieces that can move.

Unit tests updated to test for this new iterator.